fix:파일 구조 변경 및 알림 읽음여부처리
