+++ src/main/java/com/example/exodia/evalution/repository/EvalutionRepository.java
@@ -1,9 +1,8 @@
 package com.example.exodia.evalution.repository;
 
 import com.example.exodia.evalution.domain.Evalution;
-import com.example.exodia.subevalution.domain.SubEvalution;
+import com.example.exodia.evalutionFrame.subevalution.domain.SubEvalution;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
