+++ src/main/java/com/example/exodia/evalutionFrame/evalutionBig/repository/EvalutionbRepository.java
@@ -1,0 +1,9 @@
+package com.example.exodia.evalutionFrame.evalutionBig.repository;
+
+import com.example.exodia.evalutionFrame.evalutionBig.domain.Evalutionb;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface EvalutionbRepository extends JpaRepository<Evalutionb, Long> {
+}
