+++ src/main/java/com/example/exodia/evalutionFrame/evalutionMiddle/repository/EvalutionmRepository.java
@@ -1,0 +1,7 @@
+package com.example.exodia.evalutionFrame.evalutionMiddle.repository;
+
+import com.example.exodia.evalutionFrame.evalutionMiddle.domain.Evalutionm;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface EvalutionmRepository extends JpaRepository<Evalutionm, Long> {
+}
