--- src/main/java/com/example/exodia/evalutionb/dto/EvalutionbDto.java
@@ -1,19 +1,0 @@
-package com.example.exodia.evalutionb.dto;
-
-import com.example.exodia.evalutionb.domain.Evalutionb;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class EvalutionbDto {
-    private String bName; // 대분류 명
-
-    public void setBName(String bName) {
-        this.bName = bName;
-    }
-}
