--- src/main/java/com/example/exodia/evalutionb/repository/EvalutionbRepository.java
@@ -1,9 +1,0 @@
-package com.example.exodia.evalutionb.repository;
-
-import com.example.exodia.evalutionb.domain.Evalutionb;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface EvalutionbRepository extends JpaRepository<Evalutionb, Long> {
-}
