--- src/main/java/com/example/exodia/evalutionm/repository/EvalutionmRepository.java
@@ -1,7 +1,0 @@
-package com.example.exodia.evalutionm.repository;
-
-import com.example.exodia.evalutionm.domain.Evalutionm;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface EvalutionmRepository extends JpaRepository<Evalutionm, Long> {
-}
